+++ src/index.ts
@@ -1,12 +1,8 @@
-export { default as IpfsCoreService } from "src/main/ipfs-core.service.js";
-export { default as IpfsCoreModule } from "src/main/ipfs-core.module.js";
-export { default as IpfsError } from "src/errors/ipfs.error.js";
-export { default as IpfsMetricsInterceptor } from "src/interceptors/ipfs-metrics.interceptor.js";
-export type {
-	default as IpfsConfig,
-	RedisConfig,
-	PrometheusConfig,
-} from "src/types/ipfs-config.js";
-export type { default as IpfsCoreServiceDto, FileMetadata } from "src/types/ipfs-core.interface.js";
-export { IpfsErrorType } from "src/types/ipfs-error.types.js";
-export type { IpfsErrorCause } from "src/types/ipfs-error.types.js";
+export { default as IpfsCoreService } from "src/main/ipfs-core.service";
+export { default as IpfsCoreModule } from "src/main/ipfs-core.module";
+export { default as IpfsError } from "src/errors/ipfs.error";
+export { default as IpfsMetricsInterceptor } from "src/interceptors/ipfs-metrics.interceptor";
+export type { default as IpfsConfig, RedisConfig, PrometheusConfig } from "src/types/ipfs-config";
+export type { default as IpfsCoreServiceDto, FileMetadata } from "src/types/ipfs-core.interface";
+export { IpfsErrorType } from "src/types/ipfs-error.types";
+export type { IpfsErrorCause } from "src/types/ipfs-error.types";